refactor: client, controller, errorview 제거
