--- src/main/java/lotto/client/ErrorClient.java
@@ -1,5 +1,0 @@
-package lotto.client;
-
-public interface ErrorClient {
-	void error(String message);
-}
