--- src/main/java/lotto/client/InputClient.java
@@ -1,9 +1,0 @@
-package lotto.client;
-
-import java.util.List;
-
-public interface InputClient {
-    String input(String prompt);
-
-    List<String> inputRepeat(String prompt, int count);
-}
