--- src/main/java/lotto/client/OutputClient.java
@@ -1,5 +1,0 @@
-package lotto.client;
-
-public interface OutputClient {
-    void output(String message);
-}
