+++ src/main/java/lotto/view/InputView.java
@@ -1,40 +1,53 @@
 package lotto.view;
 
 import java.util.List;
+import java.util.Scanner;
+import java.util.stream.Collectors;
+import java.util.stream.IntStream;
 
-import lotto.client.ConsoleInputClient;
-import lotto.client.InputClient;
+import lotto.utils.IntegerUtils;
 
 public class InputView {
 
-    private final InputClient client;
+    private static final Scanner scanner = new Scanner(System.in);
 
-    private InputView(InputClient client) {
-        this.client = client;
+    private InputView() {
     }
 
     private static class InputViewHelper {
-        private static final InputView INSTANCE = new InputView(ConsoleInputClient.getInstance());
+        private static final InputView INSTANCE = new InputView();
     }
 
     public static InputView getInstance() {
         return InputViewHelper.INSTANCE;
     }
 
+    private String input(String prompt) {
+        System.out.print(prompt);
+        return scanner.nextLine();
+    }
+
+    private List<String> input(String prompt, int repeat) {
+        System.out.print(prompt);
+        return IntStream.range(0, repeat)
+            .mapToObj(it -> scanner.nextLine())
+            .collect(Collectors.toList());
+    }
+
     public String inputMoney() {
-        return client.input("구입금액을 입력해 주세요.\n");
+        return input("구입금액을 입력해 주세요.\n");
     }
 
     public List<String> inputLottoNumbers() {
-        int lottoCount = Integer.parseInt(client.input("수동으로 구매할 로또 수를 입력해 주세요.\n"));
-        return client.inputRepeat("수동으로 구매할 번호를 입력해 주세요.\n", lottoCount);
+        int lottoCount = IntegerUtils.parse(input("수동으로 구매할 로또 수를 입력해 주세요.\n"));
+        return input("수동으로 구매할 번호를 입력해 주세요.\n", lottoCount);
     }
 
     public String inputWinningNumber() {
-        return client.input("지난 주 당첨 번호를 입력해 주세요.\n");
+        return input("지난 주 당첨 번호를 입력해 주세요.\n");
     }
 
     public String inputBonusBall() {
-        return client.input("보너스 볼을 입력해 주세요.\n");
+        return input("보너스 볼을 입력해 주세요.\n");
     }
 }
