+++ src/main/java/lotto/view/OutputView.java
@@ -3,38 +3,37 @@ package lotto.view;
 import java.util.List;
 import java.util.stream.Collectors;
 
-import lotto.client.ConsoleOutputClient;
-import lotto.client.OutputClient;
 import lotto.dto.LottoStatisticsResponse;
 import lotto.dto.LottoTicketResponse;
 import lotto.dto.LottoWinningResponse;
 
 public class OutputView {
 
-    private final OutputClient client;
-
-    private OutputView(OutputClient client) {
-        this.client = client;
+    private OutputView() {
     }
 
     private static class OutputViewHelper {
-        private static final OutputView INSTANCE = new OutputView(ConsoleOutputClient.getInstance());
+        private static final OutputView INSTANCE = new OutputView();
     }
 
     public static OutputView getInstance() {
         return OutputViewHelper.INSTANCE;
     }
 
+    public void outputError(Exception exception) {
+        System.out.println(exception.getMessage());
+    }
+
     public void outputTickets(List<LottoTicketResponse> manualResponses, List<LottoTicketResponse> autoResponses) {
-        client.output(String.format(
+        System.out.printf(
             "수동으로 %d개, 자동으로 %d개를 구매했습니다.\n", manualResponses.size(), autoResponses.size()
-        ));
+        );
         manualResponses.forEach(this::outputTicket);
         autoResponses.forEach(this::outputTicket);
     }
 
     private void outputTicket(LottoTicketResponse response) {
-        client.output(sortTicket(response).toString() + "\n");
+        System.out.print(sortTicket(response).toString() + "\n");
     }
 
     private List<Integer> sortTicket(LottoTicketResponse response) {
@@ -50,15 +49,15 @@ public class OutputView {
     }
 
     private void outputPrompt() {
-        client.output("당첨 통계\n");
-        client.output("---------\n");
+        System.out.print("당첨 통계\n");
+        System.out.print("---------\n");
     }
 
     private void outputMatches(LottoStatisticsResponse statistics) {
         List<LottoWinningResponse> responses = statistics.getWinningResponses();
         responses.stream()
             .sorted()
-            .forEach(response -> client.output(formatWinningResponse(response) + "\n"));
+            .forEach(response -> System.out.print(formatWinningResponse(response) + "\n"));
     }
 
     private String formatWinningResponse(LottoWinningResponse response) {
@@ -70,6 +69,6 @@ public class OutputView {
     }
 
     private void outputProfit(LottoStatisticsResponse statistics) {
-        client.output(String.format("총 수익률은 %.2f입니다.\n", statistics.getProfitRate()));
+        System.out.printf("총 수익률은 %.2f입니다.\n", statistics.getProfitRate());
     }
 }
